Sources/Python/requirements.txt
# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pillow==10.1.0

# ML & Computer Vision
torch==2.1.1
torchvision==0.16.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@main
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78

# Data processing
numpy==1.24.3
scipy==1.11.4
pandas==2.1.3

# Core ML conversion
coremltools==7.0

# Utilities
tqdm==4.66.1
pyyaml==6.0.1
python-dotenv==1.0.0
